Money is now stored as a float instead of int to match TowerData cost

+++ Assets/000_TowerDefense/Scripts/Global/Globals.cs
@@ -8,8 +8,8 @@ namespace LGDP.TowerDefense
     public static class Globals
     {
 
-        private static int _money = 500;
-        public static int Money {
+        private static float _money = 500;
+        public static float Money {
             get => _money;
             set
             {
